eliminate useLiveDataForAssetKeys
