+++ js_modules/dagster-ui/packages/ui-core/src/asset-graph/Utils.tsx
@@ -2,10 +2,6 @@ import {pathHorizontalDiagonal, pathVerticalDiagonal} from '@vx/shape';
 
 import {featureEnabled, FeatureFlag} from '../app/Flags';
 import {COMMON_COLLATOR} from '../app/Util';
-import {RunStatus, StaleStatus} from '../graphql/types';
-
-import {AssetNodeKeyFragment} from './types/AssetNode.types';
-import {AssetNodeForGraphQueryFragment} from './types/useAssetGraphData.types';
 import {
   AssetGraphLiveQuery,
   AssetLatestInfoFragment,
@@ -14,7 +10,11 @@ import {
   AssetNodeLiveFreshnessInfoFragment,
   AssetNodeLiveMaterializationFragment,
   AssetNodeLiveObservationFragment,
-} from './types/useLiveDataForAssetKeys.types';
+} from '../asset-data/types/AssetLiveDataProvider.types';
+import {RunStatus, StaleStatus} from '../graphql/types';
+
+import {AssetNodeKeyFragment} from './types/AssetNode.types';
+import {AssetNodeForGraphQueryFragment} from './types/useAssetGraphData.types';
 
 type AssetNode = AssetNodeForGraphQueryFragment;
 type AssetKey = AssetNodeKeyFragment;
