+++ js_modules/dagster-ui/packages/ui-core/src/assets/FailedRunSinceMaterializationBanner.tsx
@@ -7,7 +7,7 @@ import {
 import React from 'react';
 import {Link} from 'react-router-dom';
 
-import {AssetLatestInfoRunFragment} from '../asset-graph/types/useLiveDataForAssetKeys.types';
+import {AssetLatestInfoRunFragment} from '../asset-data/types/AssetLiveDataProvider.types';
 import {titleForRun} from '../runs/RunUtils';
 import {useStepLogs} from '../runs/StepLogsDialog';
