+++ js_modules/dagster-ui/packages/ui-core/src/assets/__fixtures__/AssetTables.fixtures.ts
@@ -1,8 +1,8 @@
 import {MockedResponse} from '@apollo/client/testing';
 
+import {ASSETS_GRAPH_LIVE_QUERY} from '../../asset-data/AssetLiveDataProvider';
+import {AssetGraphLiveQuery} from '../../asset-data/types/AssetLiveDataProvider.types';
 import {MockStaleReasonData} from '../../asset-graph/__fixtures__/AssetNode.fixtures';
-import {AssetGraphLiveQuery} from '../../asset-graph/types/useLiveDataForAssetKeys.types';
-import {ASSETS_GRAPH_LIVE_QUERY} from '../../asset-graph/useLiveDataForAssetKeys';
 import {
   StaleStatus,
   RunStatus,
